feat(collection-button): rename `category-selector` to `collection-button`
