+++ components/index.js
@@ -23,12 +23,12 @@ export { default as CartSidebar } from './cart-sidebar';
 export { default as CartSummary } from './cart-summary';
 export { default as CartTotals } from './cart-totals';
 export { default as CartUpsellMenu } from './cart-upsell-menu';
-export { default as CategorySelector } from './category-selector.vue';
 export { default as CheckoutAuthentication } from './checkout-authentication';
 export { default as CheckoutItemCard } from './checkout-item-card';
 export { default as CleanseGuide } from './cleanse-guide';
 export { default as CollapsibleDrawer } from './collapsible-drawer';
 export { default as CollapsibleWalletDrawer } from './collapsible-wallet-drawer';
+export { default as CollectionButton } from './collection-button.vue';
 export { default as ConfirmationDialogTemplate } from './confirmation-dialog-template';
 export { default as ConfirmDrawer } from './confirm-drawer';
 export { default as ContentWrapper } from './content-wrapper';
