+++ snapshots.js
@@ -15,8 +15,8 @@ const groups = [{
 		'ButtonBar',
 		'Buttons',
 		'Card',
-		'CategorySelector',
 		'CollapsibleDrawer',
+		'CollectionButton',
 		'ConfirmationDialogTemplate',
 		'ConfirmDrawer',
 		'Drawer',
